+++ PicPi/Navigator.swift
@@ -11,33 +11,3 @@ import Photos
 protocol Navigator {
   func presentAddPhotoFlow(preselectedAssets: [PHAsset], preselectedFrames: [PictureFrame])
 }
-
-func injectNavigator() -> Navigator {
-  return NavigatorImpl.shared
-}
-
-class NavigatorImpl: Navigator {
-  fileprivate static let shared = NavigatorImpl()
-
-  private var appDelegate: AppDelegate {
-    UIApplication.shared.delegate as! AppDelegate
-  }
-
-  private var topViewController: UIViewController {
-    guard let rootVc = appDelegate.window!.rootViewController else {
-      fatalError("Root window has not root Vc")
-    }
-
-    return rootVc
-  }
-
-  func presentAddPhotoFlow(preselectedAssets: [PHAsset], preselectedFrames: [PictureFrame]) {
-    let addPhotoVc = AddPhotoViewController(
-      selectedAssets: preselectedAssets,
-      selectedPictureFrames: preselectedFrames
-    )
-    let navVc = UINavigationController(rootViewController: addPhotoVc)
-    navVc.modalPresentationStyle = .fullScreen
-    topViewController.present(navVc, animated: true)
-  }
-}
